Updated translations for categories (#39)
