+++ public/locales/en/translation.json
@@ -117,16 +117,28 @@
     "categories": "Categories"
   },
   "categories": {
-    "environment": "Environment",
     "animals": "Animals",
-    "children": "Children",
-    "seniors": "Seniors",
-    "women": "Women",
-    "nature": "Nature",
+    "children": "Children & Youth",
+    "community": "Community",
+    "education & literacy": "Education & Literacy",
+    "emergency & safety": "Emergency & Safety",
+    "employment": "Employment",
+    "environment": "Environment",
+    "health & medicine": "Health & Medicine",
     "homeless": "Homeless",
-    "education": "Education & Literacy",
     "hunger": "Hunger",
-    "youth": "Children & Youth"
+    "nature": "Nature",
+    "poverty": "Poverty",
+    "programming": "Programming",
+    "seniors": "Seniors",
+    "education": "Education & Literacy",
+    "youth": "Children & Youth",
+    "science": "Science",
+    "special needs": "Special needs",
+    "sports & recreation": "Sports & Recreation",
+    "students": "Students",
+    "technology": "Technology",
+    "women": "Women"
   },
   "teaminfo": {
     "members": "Members",
