+++ src/www/containers/actionprofile/action.json
@@ -4,8 +4,7 @@
   "categories": [
     { "name": "Programming" },
     { "name": "Students" },
-    { "name": "Technology" },
-    { "name": "Something" }
+    { "name": "Technology" }
   ],
   "location": {
     "name": "Εθνικής Αντίστασης 16, Καλαμαριά",
