+++ src/www/containers/actionspage/actions.json
@@ -21,7 +21,7 @@
     "description": "The Parent Support & Training Program at Easter Seals Central Texas is offering a Nurturing Parenting Group to support parents through meaningful connection and learning from an evidence-based curriculum. This group will meet every Thursday from 10am to 12pm from March 26 to May 14. In order to make this group a success, we need your help to provide childcare! Caregivers will need to be comfortable providing care to children with a variety of special needs. Childcare happens in our indoor playground room, and we provide plenty of toys and activities for the children. We will ask for all childcare providers to complete a background check. Previous volunteers say that this volunteer opportunity is fun and meaningful!",
     "date": "Saturday, 25 July 2020, 12:00",
     "location": "8505 Cross Park Drive, Suite 120, Austin, TX 78754",
-    "categories": ["Children & Youth", "Special Needs"]
+    "categories": ["Children", "Special Needs"]
   },
   "3": {
     "name": "Plant Walk",
@@ -45,7 +45,7 @@
     "description": "AVANCE-Austin serve at-risk families through it's dual-generational Parent-Child Education Program. We will be celebrating the graudation of the famlies, with chidren ages 0-3 years old, from our intensive nine month program which is focused on developing parent-child relationships through parenting classes and childhood development education. Families will be receiving a special diploma during AVANCE's graduation ceremony that is conducted in Spanish.",
     "date": "Saturday, 3 October 2020, 16:00",
     "location": "Riverside, CA 92501",
-    "categories": ["Children & Youth", "Education & Literacy", "Poverty"]
+    "categories": ["Children", "Education & Literacy", "Poverty"]
   },
   "6": {
     "name": "Share your knowledge with a small business owner",
