+++ src/www/containers/actionspage/categories.json
@@ -3,7 +3,7 @@
     "name": "Animals"
   },
   "Category2": {
-    "name": "Children & Youth"
+    "name": "Children"
   },
   "Category3": {
     "name": "Community"
