+++ src/www/containers/teamspage/teams.json
@@ -27,13 +27,13 @@
     "name": "UNICEF",
     "description": "UNICEF works in over 190 countries and territories to save children’s lives, to defend their rights, and to help them fulfil their potential, from early childhood through adolescence. And we never give up.",
     "logo": "/img/fakedata/unicef.png",
-    "categories": ["Children & Youth", "Community"]
+    "categories": ["Children", "Community"]
   },
   "team6": {
     "name": "Το Χαμόγελο του Παιδιού",
     "description": "«Το Χαμόγελο του Παιδιού» ιδρύθηκε το 1995 από τον 10χρονο Ανδρέα Γιαννόπουλο, ο οποίος λίγο πριν φύγει από τη ζωή εξέφρασε στο ημερολόγιό του την επιθυμία να δημιουργηθεί ένας σύλλογος για να έχουν όλα τα παιδιά όσα ο ίδιος είχε απλόχερα: αγάπη, στοργή, ενδιαφέρον, σεβασμό. Το όνειρο του 10χρονου Ανδρέα έγινε πραγματικότητα και σήμερα ο Οργανισμός «Το Χαμόγελο του Παιδιού», έχοντας αναγνωριστεί διεθνώς, με όραμα: το Χαμόγελο κάθε παιδιού, υλοποιεί πανελλαδικά, 365 ημέρες το χρόνο, 24 ώρες την ημέρα, αποτελεσματικές και άμεσες δράσεις για την πρόληψη και αντιμετώπιση συγκεκριμένων φαινομένων που απειλούν τα παιδιά. Μέχρι σήμερα «Το Χαμόγελο του Παιδιού» έχει στηρίξει περισσότερα από 1.200.000 παιδιά και τις οικογένειές τους. Παιδιά θύματα κάθε μορφής Βίας, Παιδιά θύματα Εξαφάνισης, Παιδιά με προβλήματα Υγείας, Παιδιά που βρίσκονται ή απειλούνται να βρεθούν σε κατάσταση φτώχειας, έχουν βρει λύση. Ο Οργανισμός στελεχώνεται από εξειδικευμένο επιστημονικό προσωπικό με την πολύτιμη υποστήριξη εθελοντών, ενώ στηρίζεται στους πολίτες και τις εταιρείες, οι οποίοι εμπιστευόμενοι το έργο του, το στηρίζουν έμπρακτα μέσω δωρεών και χορηγιών. Τα οικονομικά στοιχεία, τα οποία και δημοσιοποιούνται, ελέγχονται από το Διεθνές Σώμα Ορκωτών Ελεγκτών Ernst and Young.",
     "logo": "/img/fakedata/kids_smile.png",
-    "categories": ["Children & Youth"]
+    "categories": ["Children"]
   },
   "team7": {
     "name": "Ίδρυμα Σταύρος Νιάρχος",
