+++ src/www/containers/userprofile/actionsAttended.json
@@ -9,7 +9,7 @@
     "name": "Volunteer for AVANCE Graduation",
     "description": "AVANCE-Austin serve at-risk families through it's dual-generational Parent-Child Education Program. We will be celebrating the graudation of the famlies, with chidren ages 0-3 years old, from our intensive nine month program which is focused on developing parent-child relationships through parenting classes and childhood development education. Families will be receiving a special diploma during AVANCE's graduation ceremony that is conducted in Spanish.",
     "categories": [
-      { "name": "Children & Youth" },
+      { "name": "Children" },
       { "name": "Education & Literacy" },
       { "name": "Poverty" }
     ],
