+++ src/www/containers/userprofile/actionsOrganized.json
@@ -14,7 +14,7 @@
   "Action3": {
     "name": "Caregivers for Children with Special Needs - Nurturing Parenting Group!",
     "description": "The Parent Support & Training Program at Easter Seals Central Texas is offering a Nurturing Parenting Group to support parents through meaningful connection and learning from an evidence-based curriculum. This group will meet every Thursday from 10am to 12pm from March 26 to May 14. In order to make this group a success, we need your help to provide childcare! Caregivers will need to be comfortable providing care to children with a variety of special needs. Childcare happens in our indoor playground room, and we provide plenty of toys and activities for the children. We will ask for all childcare providers to complete a background check. Previous volunteers say that this volunteer opportunity is fun and meaningful!",
-    "categories": [{ "name": "Children & Youth" }, { "name": "Special Needs" }],
+    "categories": [{ "name": "Children" }, { "name": "Special Needs" }],
     "photo": "/img/fakedata/Beck.jpg"
   },
   "Action4": {
