+++ src/www/containers/userprofile/teamsFollow.json
@@ -8,7 +8,7 @@
   "Team2": {
     "name": "Το Χαμόγελο του Παιδιού",
     "description": "«Το Χαμόγελο του Παιδιού» ιδρύθηκε το 1995 από τον 10χρονο Ανδρέα Γιαννόπουλο, ο οποίος λίγο πριν φύγει από τη ζωή εξέφρασε στο ημερολόγιό του την επιθυμία να δημιουργηθεί ένας σύλλογος για να έχουν όλα τα παιδιά όσα ο ίδιος είχε απλόχερα: αγάπη, στοργή, ενδιαφέρον, σεβασμό. Το όνειρο του 10χρονου Ανδρέα έγινε πραγματικότητα και σήμερα ο Οργανισμός «Το Χαμόγελο του Παιδιού», έχοντας αναγνωριστεί διεθνώς, με όραμα: το Χαμόγελο κάθε παιδιού, υλοποιεί πανελλαδικά, 365 ημέρες το χρόνο, 24 ώρες την ημέρα, αποτελεσματικές και άμεσες δράσεις για την πρόληψη και αντιμετώπιση συγκεκριμένων φαινομένων που απειλούν τα παιδιά. Μέχρι σήμερα «Το Χαμόγελο του Παιδιού» έχει στηρίξει περισσότερα από 1.200.000 παιδιά και τις οικογένειές τους. Παιδιά θύματα κάθε μορφής Βίας, Παιδιά θύματα Εξαφάνισης, Παιδιά με προβλήματα Υγείας, Παιδιά που βρίσκονται ή απειλούνται να βρεθούν σε κατάσταση φτώχειας, έχουν βρει λύση. Ο Οργανισμός στελεχώνεται από εξειδικευμένο επιστημονικό προσωπικό με την πολύτιμη υποστήριξη εθελοντών, ενώ στηρίζεται στους πολίτες και τις εταιρείες, οι οποίοι εμπιστευόμενοι το έργο του, το στηρίζουν έμπρακτα μέσω δωρεών και χορηγιών. Τα οικονομικά στοιχεία, τα οποία και δημοσιοποιούνται, ελέγχονται από το Διεθνές Σώμα Ορκωτών Ελεγκτών Ernst and Young.",
-    "categories": [{ "name": "Children & Youth" }],
+    "categories": [{ "name": "Children" }],
     "photo": "/img/fakedata/kids_smile.png"
   }
 }
